libatari800: moved api functions into own source file, moved jmp_buf declaration into cpu_crash.h
