+++ src/libatari800/main.h
@@ -5,13 +5,6 @@
 
 #include "config.h"
 
-#ifdef HAVE_SETJMP
-#include <setjmp.h>
-extern jmp_buf libatari800_cpu_crash;
-#endif /* HAVE_SETJMP */
-
-#include "libatari800/libatari800.h"
-
 void LIBATARI800_Frame(void);
 
 #endif /* LIBATARI800_VIDEO_H_ */
